Made the methods synchronized on session objects, other minor improvements
