+++ src/test/java/com/es/phoneshop/web/ProductDetailsPageServletTest.java
@@ -4,8 +4,8 @@ import com.es.phoneshop.model.cart.Cart;
 import com.es.phoneshop.model.cart.CartService;
 import com.es.phoneshop.model.product.ArrayListProductDao;
 import com.es.phoneshop.model.product.Product;
-import com.es.phoneshop.model.recently_viewed_products.RecentlyViewedProducts;
-import com.es.phoneshop.model.recently_viewed_products.RecentlyViewedProductsService;
+import com.es.phoneshop.model.RecentlyViewedProducts.RecentlyViewedProducts;
+import com.es.phoneshop.model.RecentlyViewedProducts.RecentlyViewedProductsService;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
